+++ src/array/codec.rs
@@ -55,7 +55,6 @@ pub use bytes_to_bytes::gzip::{GzipCodec, GzipCodecConfiguration, GzipCodecConfi
 #[cfg(feature = "zstd")]
 pub use bytes_to_bytes::zstd::{ZstdCodec, ZstdCodecConfiguration, ZstdCodecConfigurationV1};
 
-use itertools::Itertools;
 use thiserror::Error;
 
 mod array_partial_decoder_cache;
@@ -69,9 +68,10 @@ pub use byte_interval_partial_decoder::ByteIntervalPartialDecoder;
 #[cfg(feature = "async")]
 pub use byte_interval_partial_decoder::AsyncByteIntervalPartialDecoder;
 
+use crate::byte_range::extract_byte_ranges_read_seek;
 use crate::{
     array_subset::{ArraySubset, IncompatibleArraySubsetAndShapeError},
-    byte_range::{ByteOffset, ByteRange, InvalidByteRangeError},
+    byte_range::{ByteRange, InvalidByteRangeError},
     metadata::v3::MetadataV3,
     plugin::{Plugin, PluginCreateError},
     storage::{ReadableStorage, StorageError, StoreKey},
@@ -82,10 +82,6 @@ use crate::storage::AsyncReadableStorage;
 
 use std::borrow::Cow;
 use std::sync::Arc;
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    io::{Read, Seek, SeekFrom},
-};
 
 use super::{
     concurrency::RecommendedConcurrency, ArrayMetadataOptions, BytesRepresentation,
@@ -786,158 +782,5 @@ impl From<&str> for CodecError {
 impl From<String> for CodecError {
     fn from(err: String) -> Self {
         Self::Other(err)
-    }
-}
-
-/// Extract byte ranges from bytes implementing [`Read`] and [`Seek`].
-///
-/// # Errors
-///
-/// Returns a [`std::io::Error`] if there is an error reading or seeking from `bytes`.
-/// This can occur if the byte range is out-of-bounds of the `bytes`.
-///
-/// # Panics
-///
-/// Panics if a byte has length exceeding [`usize::MAX`].
-pub fn extract_byte_ranges_read_seek<T: Read + Seek>(
-    bytes: &mut T,
-    byte_ranges: &[ByteRange],
-) -> std::io::Result<Vec<Vec<u8>>> {
-    let len: u64 = bytes.seek(SeekFrom::End(0))?;
-    let mut out = Vec::with_capacity(byte_ranges.len());
-    for byte_range in byte_ranges {
-        let data: Vec<u8> = match byte_range {
-            ByteRange::FromStart(offset, None) => {
-                bytes.seek(SeekFrom::Start(*offset))?;
-                let length = usize::try_from(len).unwrap();
-                let mut data = vec![0; length];
-                bytes.read_exact(&mut data)?;
-                data
-            }
-            ByteRange::FromStart(offset, Some(length)) => {
-                bytes.seek(SeekFrom::Start(*offset))?;
-                let length = usize::try_from(*length).unwrap();
-                let mut data = vec![0; length];
-                bytes.read_exact(&mut data)?;
-                data
-            }
-            ByteRange::FromEnd(offset, None) => {
-                bytes.seek(SeekFrom::Start(0))?;
-                let length = usize::try_from(len - offset).unwrap();
-                let mut data = vec![0; length];
-                bytes.read_exact(&mut data)?;
-                data
-            }
-            ByteRange::FromEnd(offset, Some(length)) => {
-                bytes.seek(SeekFrom::End(-i64::try_from(*offset + *length).unwrap()))?;
-                let length = usize::try_from(*length).unwrap();
-                let mut data = vec![0; length];
-                bytes.read_exact(&mut data)?;
-                data
-            }
-        };
-        out.push(data);
-    }
-    Ok(out)
-}
-
-/// Extract byte ranges from bytes implementing [`Read`].
-///
-/// # Errors
-///
-/// Returns a [`std::io::Error`] if there is an error reading from `bytes`.
-/// This can occur if the byte range is out-of-bounds of the `bytes`.
-///
-/// # Panics
-///
-/// Panics if a byte has length exceeding [`usize::MAX`].
-pub fn extract_byte_ranges_read<T: Read>(
-    bytes: &mut T,
-    size: u64,
-    byte_ranges: &[ByteRange],
-) -> std::io::Result<Vec<Vec<u8>>> {
-    // Could this be cleaner/more efficient?
-
-    // Allocate output and find the endpoints of the "segments" of bytes which must be read
-    let mut out = Vec::with_capacity(byte_ranges.len());
-    let mut segments_endpoints = BTreeSet::<u64>::new();
-    for byte_range in byte_ranges {
-        out.push(vec![0; usize::try_from(byte_range.length(size)).unwrap()]);
-        segments_endpoints.insert(byte_range.start(size));
-        segments_endpoints.insert(byte_range.end(size));
-    }
-
-    // Find the overlapping part of each byte range with each segment
-    //                 SEGMENT start     , end        OUTPUT index, offset
-    let mut overlap: BTreeMap<(ByteOffset, ByteOffset), Vec<(usize, ByteOffset)>> = BTreeMap::new();
-    for (byte_range_index, byte_range) in byte_ranges.iter().enumerate() {
-        let byte_range_start = byte_range.start(size);
-        let range = segments_endpoints.range((
-            std::ops::Bound::Included(byte_range_start),
-            std::ops::Bound::Included(byte_range.end(size)),
-        ));
-        for (segment_start, segment_end) in range.tuple_windows() {
-            let byte_range_offset = *segment_start - byte_range_start;
-            overlap
-                .entry((*segment_start, *segment_end))
-                .or_default()
-                .push((byte_range_index, byte_range_offset));
-        }
-    }
-
-    let mut bytes_offset = 0u64;
-    for ((segment_start, segment_end), outputs) in overlap {
-        // Go to the start of the segment
-        if segment_start > bytes_offset {
-            std::io::copy(
-                &mut bytes.take(segment_start - bytes_offset),
-                &mut std::io::sink(),
-            )
-            .unwrap();
-        }
-
-        let segment_length = segment_end - segment_start;
-        if outputs.is_empty() {
-            // No byte ranges are associated with this segment, so just read it to sink
-            std::io::copy(&mut bytes.take(segment_length), &mut std::io::sink()).unwrap();
-        } else {
-            // Populate all byte ranges in this segment with data
-            let segment_length_usize = usize::try_from(segment_length).unwrap();
-            let mut segment_bytes = vec![0; segment_length_usize];
-            bytes.take(segment_length).read_exact(&mut segment_bytes)?;
-            for (byte_range_index, byte_range_offset) in outputs {
-                let byte_range_offset = usize::try_from(byte_range_offset).unwrap();
-                out[byte_range_index][byte_range_offset..byte_range_offset + segment_length_usize]
-                    .copy_from_slice(&segment_bytes);
-            }
-        }
-
-        // Offset is now the end of the segment
-        bytes_offset = segment_end;
-    }
-
-    Ok(out)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_extract_byte_ranges_read() {
-        let data: Vec<u8> = (0..10).collect();
-        let size = data.len() as u64;
-        let mut read = std::io::Cursor::new(data);
-        let byte_ranges = vec![
-            ByteRange::FromStart(3, Some(3)),
-            ByteRange::FromStart(4, Some(1)),
-            ByteRange::FromStart(1, Some(1)),
-            ByteRange::FromEnd(1, Some(5)),
-        ];
-        let out = extract_byte_ranges_read(&mut read, size, &byte_ranges).unwrap();
-        assert_eq!(
-            out,
-            vec![vec![3, 4, 5], vec![4], vec![1], vec![4, 5, 6, 7, 8]]
-        );
     }
 }